optimization workflow component creation
