+++ src/stubs/angular/make-angular-component.mts
@@ -1,0 +1,25 @@
+import * as fs from 'fs';
+import { writeFile } from '../../utils/utils.mjs';
+
+export function makeAngularComponent(filePathDestination: string, templateFilePath: string, componentName: string): void {
+	fs.readFile(templateFilePath, 'utf8', (err, component) => {
+
+		component = component.replace(/selector: 'SelectorName'/, `selector: 'app-${convertFromCamelCase(componentName)}'`);
+		component = component.replace(/ComponentName/g, `${convertToCamelCase(componentName)}Component`);
+		writeFile(filePathDestination, component);
+
+	});
+
+}
+
+function convertToCamelCase(string: string): string {
+	return string.replace(/-([a-z])/g, (s: string) => {
+		return s.toUpperCase();
+	}).replace(/^[a-z]/, (s) => {
+		return s.toUpperCase();
+	});
+}
+
+function convertFromCamelCase(string: string): string {
+	return string.replace(/([a-z])([A-Z])/g, '$1-$2').toLowerCase();
+}
